+++ R/board_azure.R
@@ -113,7 +113,7 @@ pin_fetch.pins_board_azure <- function(board, name, version = NULL, ...) {
 pin_store.pins_board_azure <- function(board, name, paths, metadata,
                                        versioned = NULL, ...) {
   check_name(name)
-  version <- version_setup(board, name, metadata, versioned = versioned)
+  version <- version_setup(board, name, version_name(metadata), versioned = versioned)
 
   version_dir <- fs::path(name, version)