+++ R/meta.R
@@ -54,6 +54,12 @@ parse_8601_compact <- function(x) {
   y
 }
 
+parse_8601 <- function(x) {
+  y <- as.POSIXct(strptime(x, "%Y-%m-%dT%H:%M", tz = "UTC"))
+  attr(y, "tzone") <- ""
+  y
+}
+
 default_title <- function(name, data = NULL, path = NULL) {
   if (!xor(is.null(data), is.null(path))) {
     abort("Must supply exactly one of `path` and `data`")